Hyper.Config.File.Parser: bugfix for site update (comments)
There was a bug that we only kept track of the default site but not the
current one we're operating on.  Due to that we could never update the
specific sites beyond the default configuration.  After adding another
entry in the state to deal with this the code also had to be changed to
properly insert the data.