Hyper.Config.File.Parser: updated p_configuration do post-parse config update (comments)
During the parse, the default site configuration is simply supplying
defaults for the other site definitions to use (e.g. root directory for
relative paths) but before parse returns, this has to be switched with
the site that was actually configured as the default.  Note that if the
default site doesn't have a seperate entry then the server values will
be used.