TemplateEngine: bug fixed problem with upDir (comments)
takeDirectory will only actually move directories if the path doesn't end in a slash.  This was causing the same path to be tried twice for every new call of applyTemplate since the first call was the file, the second call chopped the slash off the end and reapplied the same doesFileExist test on the same file and the third call finally moved up a directory.  Moved the dropTrailingPathSeperator to the upDir function