TemplateEngineConduit: major change to fix div problem (comments)
I changed everything from using Maybe to use List context.  This allows us to still support failure but it also supports a command taking one element and returning more than one.  This also fixes our problem with nodes being processed backwards because since processElement could return a list we have to append it in any case so we just put newly processed children on the end.

This appears to fix the problem of having divs around variables when not needed (and could make them not usable in many places like inside a title tag)